requirements.txt
streamlit
openai
markdown
reportlab
matplotlib
numpy
pillow
google-genai
pandas
